+++ utilities/on_subscibe-service/dotnet/Program.cs
@@ -41,27 +41,4 @@ app.MapGet("/health", () => Results.Ok("Utility is running"))
 
 app.Run();
 
-public static class Handlers
-{
-    public static IResult GenerateKeysHandler()
-    {
-        try
-        {
-            var (signingPublicKey, signingPrivateKey) = ("key1", "key2");
-            var (encPublicKey, encPrivateKey) = ("key3", "key4");
-
-            return Results.Ok(new
-            {
-                signing_public_key = signingPublicKey,
-                signing_private_key = signingPrivateKey,
-                encryption_public_key = encPublicKey,
-                encryption_private_key = encPrivateKey
-            });
-        }
-        catch (Exception ex)
-        {
-            return Results.Problem($"Error generating keys: {ex.Message}");
-        }
-    }
-}
